+++ api-tests/relationships/crud/one-to-many.test.js
@@ -130,7 +130,7 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
 
     [
       [createListsLR, 'Left -> Right'],
-      // [createListsRL, 'Right -> Left'],
+      [createListsRL, 'Right -> Left'],
     ].forEach(([createLists, order]) => {
       describe(`One-to-many relationships - ${order}`, () => {
         function setupKeystone(adapterName) {
@@ -441,13 +441,12 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
             })
           );
 
-          test.only(
+          test(
             'With create',
             runner(setupKeystone, async ({ keystone }) => {
               const { companies } = await createInitialData(keystone);
               let company = companies[0];
-              const locationName = 'Something more humble'; //sampleOne(alphanumGenerator);
-              console.log('=========================================');
+              const locationName = sampleOne(alphanumGenerator);
               const { data, errors } = await graphqlRequest({
                 keystone,
                 query: `
@@ -459,10 +458,6 @@ multiAdapterRunners().map(({ runner, adapterName }) =>
                 }
             `,
               });
-
-              console.log(keystone.adapters.JSONAdapter.logDB());
-              console.log({ updateCompanyLocations: data.updateCompany.locations });
-
               expect(errors).toBe(undefined);
 
               const { Company, Location } = await getCompanyAndLocation(