+++ packages/adapter-knex/lib/adapter-knex.js
@@ -311,6 +311,22 @@ class KnexListAdapter extends BaseListAdapter {
     );
   }
 
+  async _unsetForeignOneToOneValues(data, id) {
+    // If there's a 1:1 FK in the data on a different list we need to go and
+    // delete it from any other item;
+    await Promise.all(
+      Object.keys(data)
+        .map(key => ({ adapter: this.fieldAdaptersByPath[key] }))
+        .filter(({ adapter }) => adapter && adapter.isRelationship)
+        .filter(
+          ({ adapter: { rel } }) => rel.cardinality === '1:1' && rel.tableName !== this.tableName
+        )
+        .map(({ adapter: { rel: { tableName, columnName } } }) =>
+          this._setNullByValue({ tableName, columnName, value: id })
+        )
+    );
+  }
+
   async _processNonRealFields(data, processFunction) {
     return resolveAllKeys(
       arrayToObject(
@@ -381,6 +397,7 @@ class KnexListAdapter extends BaseListAdapter {
             .returning('id');
         }
       } else {
+        console.log('disconnect');
         return [];
       }
     }
@@ -411,7 +428,10 @@ class KnexListAdapter extends BaseListAdapter {
     const realData = pick(data, this.realKeys);
 
     // Unset any real 1:1 fields
+    console.log({ realData, data, id });
+
     await this._unsetOneToOneValues(realData);
+    await this._unsetForeignOneToOneValues(data, id);
 
     // Update the real data
     const query = this._query()
@@ -458,6 +478,8 @@ class KnexListAdapter extends BaseListAdapter {
               .whereIn(selectCol, needsDelete) // far side
               .del();
           } else {
+            console.log(tableName, needsDelete);
+
             await this._query()
               .table(tableName)
               .whereIn(selectCol, needsDelete)
@@ -544,7 +566,6 @@ class KnexListAdapter extends BaseListAdapter {
       count = Math.max(0, count); // Don't want to go negative from a skip!
       return { count };
     }
-
     return results;
   }
 }
@@ -662,29 +683,32 @@ class QueryBuilder {
     // Insert joins to handle 1:1 relationships where the FK is stored on the other table.
     // We join against the other table and select its ID as the path name, so that it appears
     // as if it existed on the primary table all along!
-    if (!meta) {
-      listAdapter.fieldAdapters
-        .filter(a => a.isRelationship && a.rel.cardinality === '1:1' && a.rel.right === a.field)
-        .forEach(a => {
-          const { tableName, columnName } = a.rel;
-          const otherTableAlias = `${tableAlias}__${a.path}_11`;
-          if (!this._tableAliases[otherTableAlias]) {
-            this._tableAliases[otherTableAlias] = true;
-            // LEFT OUTERJOIN on ... table>.<id> = <otherTable>.<columnName> SELECT <othertable>.<id> as <path>
-            query
-              .leftOuterJoin(
-                `${tableName} as ${otherTableAlias}`,
-                `${otherTableAlias}.${columnName}`,
-                `${tableAlias}.id`
-              )
-              .select(`${otherTableAlias}.id as ${a.path}`);
-          }
-        });
-    }
 
     const joinPaths = Object.keys(where).filter(
       path => !this._getQueryConditionByPath(listAdapter, path)
     );
+
+    const joinedPaths = [];
+    listAdapter.fieldAdapters
+      .filter(a => a.isRelationship && a.rel.cardinality === '1:1' && a.rel.right === a.field)
+      .forEach(({ path, rel }) => {
+        const { tableName, columnName } = rel;
+        const otherTableAlias = `${tableAlias}__${path}`;
+        if (!this._tableAliases[otherTableAlias] && (!meta || joinPaths.includes(path))) {
+          this._tableAliases[otherTableAlias] = true;
+          // LEFT OUTERJOIN on ... table>.<id> = <otherTable>.<columnName> SELECT <othertable>.<id> as <path>
+          query.leftOuterJoin(
+            `${tableName} as ${otherTableAlias}`,
+            `${otherTableAlias}.${columnName}`,
+            `${tableAlias}.id`
+          );
+          if (!meta) {
+            query.select(`${otherTableAlias}.id as ${path}`);
+          }
+          joinedPaths.push(path);
+        }
+      });
+
     for (let path of joinPaths) {
       if (path === 'AND' || path === 'OR') {
         // AND/OR we need to traverse their children
@@ -693,7 +717,7 @@ class QueryBuilder {
         const otherAdapter = listAdapter.fieldAdaptersByPath[path];
         // If no adapter is found, it must be a query of the form `foo_some`, `foo_every`, etc.
         // These correspond to many-relationships, which are handled separately
-        if (otherAdapter) {
+        if (otherAdapter && !joinedPaths.includes(path)) {
           // We need a join of the form:
           // ... LEFT OUTER JOIN {otherList} AS t1 ON {tableAlias}.{path} = t1.id
           // Each table has a unique path to the root table via foreign keys