+++ packages/fields/src/types/Text/views/Controller.js
@@ -2,7 +2,7 @@ import FieldController from '../../../Controller';
 
 export default class TextController extends FieldController {
   getFilterGraphQL = ({ type, value }) => {
-    const key = type === 'is' ? `${this.path}` : `${this.path}_${type}`;
+    const key = type === 'is_i' ? `${this.path}_i` : `${this.path}_${type}`;
     return `${key}: "${value}"`;
   };
   getFilterLabel = ({ label }) => {