+++ packages/keystone/lib/Keystone/index.js
@@ -134,6 +134,7 @@ module.exports = class Keystone {
         )
       );
     }
+
     return graphql(schema, requestString, rootValue, contextValue, variableValues, operationName);
   }
 
@@ -636,6 +637,13 @@ module.exports = class Keystone {
         schemaName: this._schemaNames.length === 1 ? this._schemaNames[0] : undefined,
       })
     );
+
+    // These function can't be called after prepare(), so make them throw an error from now on.
+    ['extendGraphQLSchema', 'createList', 'createAuthStrategy'].forEach(f => {
+      this[f] = () => {
+        throw new Error(`keystone.${f} must be called before keystone.prepare()`);
+      };
+    });
 
     return { middlewares };
   }